+++ Core/Serializer/Binary/StructuredBinaryStreamReader.h
@@ -136,7 +136,7 @@ namespace Lunia {
 
 				//dataSize-=sizeof(uint32)*3; // begin header - type hash, maj ver, min ver
 
-				while (levelStartPosition.top() + dataSize > stream->GetReadCursor())
+				while (levelStartPosition.top() + dataSize > (uint32)stream->GetReadCursor())
 				{
 					/*if (stream->GetSizeLeft() < 2*sizeof(uint32)) {
 						stream->SetReadCursor(lastPosition, IStream::CursorPosition::Begin);
@@ -168,7 +168,7 @@ namespace Lunia {
 				int lastPosition = stream->GetReadCursor();
 				//dataSize-=sizeof(int)*3; // begin header - type hash, maj ver, min ver
 
-				while (levelStartPosition.top() + dataSize > stream->GetReadCursor())
+				while (levelStartPosition.top() + dataSize > (uint32)stream->GetReadCursor())
 				{
 					/*if (stream->GetSizeLeft() < 2*sizeof(uint32)) {
 						stream->SetReadCursor(lastPosition, IStream::CursorPosition::Begin);
